# Conflicts:
#	.idea/workspace.xml

+++ Code/Controller/ControllerParametre_Classe.php
@@ -18,8 +18,10 @@ require_once __DIR__."/../Model/BDDManager.php";
 $titrefichier = "Classe";
 $stylecss = "Parametre.css";
 $titreparametre = "Classe";
+$stylecss3 = "Bouton.css";
 $bdd = initialiseConnexionBDD();
 $Message = "";
+$verif = false;
 
 if (unserialize($_SESSION['utilisateur']) instanceof Administrateur) {
     $claDAO = new ClasseDAO($bdd);
@@ -32,6 +34,7 @@ if (unserialize($_SESSION['utilisateur']) instanceof Administrateur) {
             $cla = new Classe(0, $nom, $nb);
             if ($claDAO->create($cla)) {
                 $Message = "creation reussie";
+                $verif = true;
                 $clas = $claDAO->getAll();
             } else {
                 $Message = "erreur de creation";
@@ -54,6 +57,8 @@ if (unserialize($_SESSION['utilisateur']) instanceof Administrateur) {
             } else {
                 if ($claDAO->update($cla)) {
                     $Message = "modification reussie";
+                    $verif = true;
+                    $clas = $claDAO->getAll();
                 }
             }
         } else {
@@ -63,12 +68,17 @@ if (unserialize($_SESSION['utilisateur']) instanceof Administrateur) {
 
     if (isset($_POST['btnDelete'])) {
         if (isset($_POST['classe-select'])) {
-            $cla = $claDAO->find($_POST['classe-select']);
-            if ($claDAO->delete($cla)) {
-                $Message = "suppression reussie";
-                $clas = $claDAO->getAll();
+            if (!empty($_POST['classe-select'])) {
+                $cla = $claDAO->find($_POST['classe-select']);
+                if ($claDAO->delete($cla)) {
+                    $Message = "suppression reussie";
+                    $verif = true;
+                    $clas = $claDAO->getAll();
+                } else {
+                    $Message = "Il y des etudiants dans cette classe";
+                }
             } else {
-                $Message = "Il y des etudiants dans cette classe";
+                $Message = "selectionnez une classe";
             }
         } else {
             $Message = "selectionnez une classe";
